requirements.txt
python-telegram-bot>=20.0
openai
beautifulsoup4
requests
python-dotenv
